backend/requirements.txt
boto3==1.39.4
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.176.0
pytest==8.4.1
python-lambda-local==0.1.13
python-jose==3.5.0
requests==2.32.4
Pillow==10.4.0
aws-lambda-powertools==2.43.0
stripe==11.4.0